+++ apps/meteor/client/importPackages.ts
@@ -26,4 +26,3 @@ import '../app/utils/client';
 import '../app/settings/client';
 import '../app/ui-utils/client';
 import '../app/reactions/client';
-import '../app/livechat/client';