+++ apps/meteor/client/providers/LayoutProvider.tsx
@@ -24,6 +24,7 @@ const LayoutProvider = ({ children }: LayoutProviderProps) => {
 	const breakpoints = useBreakpoints(); // ["xs", "sm", "md", "lg", "xl", xxl"]
 	const [hiddenActions, setHiddenActions] = useState(hiddenActionsDefaultValue);
 	const enhancedNavigationEnabled = useFeaturePreview('newNavigation');
+	const secondSidebarEnabled = useFeaturePreview('secondarySidebar');
 
 	const router = useRouter();
 	// Once the layout is embedded, it can't be changed
@@ -34,6 +35,7 @@ const LayoutProvider = ({ children }: LayoutProviderProps) => {
 
 	const shouldToggle = enhancedNavigationEnabled ? isTablet || isMobile : isMobile;
 	const shouldDisplaySidePanel = !isTablet || displaySidePanel;
+	const defaultSidebarWidth = secondSidebarEnabled ? '220px' : '240px';
 
 	useEffect(() => {
 		setIsCollapsed(shouldToggle);
@@ -80,7 +82,7 @@ const LayoutProvider = ({ children }: LayoutProviderProps) => {
 						openSidePanel: () => setDisplaySidePanel(true),
 					},
 					size: {
-						sidebar: isTablet ? '280px' : '240px',
+						sidebar: isTablet ? '280px' : defaultSidebarWidth,
 						// eslint-disable-next-line no-nested-ternary
 						contextualBar: breakpoints.includes('sm') ? (breakpoints.includes('xl') ? '38%' : '380px') : '100%',
 					},
@@ -100,6 +102,7 @@ const LayoutProvider = ({ children }: LayoutProviderProps) => {
 					isCollapsed,
 					shouldToggle,
 					shouldDisplaySidePanel,
+					defaultSidebarWidth,
 					breakpoints,
 					hiddenActions,
 					router,