+++ apps/meteor/tests/end-to-end/api/livechat/00-rooms.ts
@@ -22,6 +22,7 @@ import type { SuccessResult } from '../../../../app/api/server/definition';
 import { getCredentials, api, request, credentials, methodCall } from '../../../data/api-data';
 import { apps, APP_URL } from '../../../data/apps/apps-data';
 import { createCustomField } from '../../../data/livechat/custom-fields';
+import type { OnlineAgent } from '../../../data/livechat/department';
 import {
 	createDepartmentWith2OnlineAgents,
 	createDepartmentWithAnAwayAgent,
@@ -737,8 +738,7 @@ describe('LIVECHAT - rooms', () => {
 					.set(userCreds)
 					.expect(200)
 					.expect((res) => {
-						expect(res.body.rooms.length).to.be.equal(1);
-						expect(res.body.rooms[0]._id).to.be.equal(room1._id);
+						expect(res.body.rooms.some((r: IOmnichannelRoom) => r._id === room1._id)).to.be.true;
 					});
 			});
 			it('should return a valid list of rooms for monitor 2', () => {
@@ -747,22 +747,128 @@ describe('LIVECHAT - rooms', () => {
 					.set(user2Creds)
 					.expect(200)
 					.expect((res) => {
-						expect(res.body.rooms.length).to.be.equal(2);
+						expect(res.body.rooms.some((r: IOmnichannelRoom) => r._id === room2._id)).to.be.true;
+						expect(res.body.rooms.some((r: IOmnichannelRoom) => r._id === room1._id)).to.be.true;
 					});
 			});
 			it('should allow monitor 1 to filter by units', async () => {
 				const { body } = await request.get(api('livechat/rooms')).set(userCreds).query({ 'units[]': unit._id }).expect(200);
-				expect(body.rooms.length).to.be.equal(1);
-				expect(body.rooms[0]._id).to.be.equal(room1._id);
+				expect(body.rooms.some((room: IOmnichannelRoom) => room._id === room1._id)).to.be.true;
+				expect(body.rooms.find((r: IOmnichannelRoom) => r._id === room2._id)).to.be.undefined;
+				expect(body.rooms.some((r: IOmnichannelRoom) => r.departmentAncestors?.includes(unit._id))).to.be.true;
+				expect(body.rooms.every((r: IOmnichannelRoom) => !r.departmentAncestors?.includes(unit2._id))).to.be.true;
 			});
 			it('should not allow monitor 1 to filter by a unit hes not part of', async () => {
 				const { body } = await request.get(api('livechat/rooms')).set(userCreds).query({ 'units[]': unit2._id }).expect(200);
-				expect(body.rooms.length).to.be.equal(0);
+				expect(body.rooms.every((r: IOmnichannelRoom) => !r.departmentAncestors?.includes(unit2._id))).to.be.true;
 			});
 			it('should allow monitor 2 to filter by only one unit', async () => {
 				const { body } = await request.get(api('livechat/rooms')).set(user2Creds).query({ 'units[]': unit2._id }).expect(200);
+				expect(body.rooms.every((r: IOmnichannelRoom) => !r.departmentAncestors?.includes(unit._id))).to.be.true;
+				expect(body.rooms.some((r: IOmnichannelRoom) => r.departmentAncestors?.includes(unit2._id))).to.be.true;
+				expect(body.rooms.find((r: IOmnichannelRoom) => r._id === room2._id)).to.be.not.undefined;
+			});
+		});
+		(IS_EE ? describe : describe.skip)('units & room restrictions for agents & monitors', () => {
+			let user: IUser;
+			let agent: OnlineAgent;
+			let agent2: OnlineAgent;
+			let userCreds: Credentials;
+			let department: ILivechatDepartment;
+			let department2: ILivechatDepartment;
+			let unit: IOmnichannelBusinessUnit;
+			let unit2: IOmnichannelBusinessUnit;
+			let room1: IOmnichannelRoom;
+			let room2: IOmnichannelRoom;
+
+			before(async () => {
+				user = await createUser();
+				userCreds = await login(user.username!, password);
+
+				await createMonitor(user.username!);
+				const { department: dep1, agent: agent1 } = await createDepartmentWithAnOnlineAgent();
+				const { department: dep2, agent: agen2 } = await createDepartmentWithAnOnlineAgent();
+
+				department = dep1;
+				agent = agent1;
+				department2 = dep2;
+				agent2 = agen2;
+				unit = await createUnit(user._id, user.username!, [department._id]);
+				unit2 = await createUnit(user._id, user.username!, [department2._id]);
+			});
+			before(async () => {
+				const { room: localOpenRoom } = await startANewLivechatRoomAndTakeIt({
+					departmentId: department._id,
+					agent: agent.credentials,
+				});
+				const { room: localOpenRoom2 } = await startANewLivechatRoomAndTakeIt({
+					departmentId: department2._id,
+					agent: agent2.credentials,
+				});
+				room1 = localOpenRoom;
+				room2 = localOpenRoom2;
+			});
+			before(async () => {
+				await restorePermissionToRoles('view-livechat-rooms');
+			});
+			after(async () => {
+				await Promise.all([
+					deleteUser(user),
+					deleteUser(agent.user),
+					deleteUnit(unit),
+					deleteDepartment(department._id),
+					deleteUnit(unit2),
+					deleteDepartment(department2._id),
+					deleteUser(agent2.user),
+				]);
+			});
+
+			it('should show the room to the agent', async () => {
+				const { body } = await request.get(api('livechat/rooms')).query({ 'agents[]': agent.user._id }).set(agent.credentials).expect(200);
+				expect(body.rooms.length).to.be.equal(1);
+				expect(body.rooms[0]._id).to.be.equal(room1._id);
+			});
+			it('should show the closed room to the agent', async () => {
+				await closeOmnichannelRoom(room1._id);
+				const { body } = await request
+					.get(api('livechat/rooms'))
+					.query({ 'agents[]': agent.user._id, 'open': false })
+					.set(agent.credentials)
+					.expect(200);
+				expect(body.rooms.length).to.be.equal(1);
+				expect(body.rooms[0]._id).to.be.equal(room1._id);
+			});
+			it('should not show the agent a room taken by other agent', async () => {
+				const { body } = await request
+					.get(api('livechat/rooms'))
+					.query({ 'agents[]': agent.user._id, 'open': true })
+					.set(agent.credentials)
+					.expect(200);
+				expect(body.rooms.length).to.be.equal(0);
+			});
+			it('should not allow the agent to filter by another agent', async () => {
+				await request.get(api('livechat/rooms')).query({ 'agents[]': agent.user._id, 'open': true }).set(agent2.credentials).expect(403);
+			});
+			it('should allow the agent to filter its own closed room', async () => {
+				await closeOmnichannelRoom(room2._id);
+				const { body } = await request
+					.get(api('livechat/rooms'))
+					.query({ 'agents[]': agent2.user._id, 'open': false })
+					.set(agent2.credentials)
+					.expect(200);
 				expect(body.rooms.length).to.be.equal(1);
 				expect(body.rooms[0]._id).to.be.equal(room2._id);
+			});
+			it('should allow monitor to see rooms from his units', async () => {
+				const { body } = await request.get(api('livechat/rooms')).set(userCreds).query({ 'units[]': unit._id }).expect(200);
+				expect(body.rooms.find((r: IOmnichannelRoom) => r._id === room1._id)).to.be.not.undefined;
+				expect(body.rooms.find((r: IOmnichannelRoom) => r._id === room2._id)).to.be.undefined;
+			});
+			it('should show the monitor both rooms when no unit filter is applied', async () => {
+				const { body } = await request.get(api('livechat/rooms')).set(userCreds).expect(200);
+
+				expect(body.rooms.some((room: IOmnichannelRoom) => room._id === room1._id)).to.be.true;
+				expect(body.rooms.some((room: IOmnichannelRoom) => room._id === room2._id)).to.be.true;
 			});
 		});
 	});