+++ packages/mock-providers/src/MockedAppRootBuilder.tsx
@@ -556,14 +556,15 @@ export class MockedAppRootBuilder {
 
 	// Manually changes the language in the i18next instance
 	// To be used with languages other than the default one
-	withTranslationLanguage(lng: string): this {
+	withDefaultLanguage(lng: string): this {
 		if (this.i18n.isInitialized) {
 			this.i18n.changeLanguage(lng);
-		} else {
-			this.i18n.on('initialized', () => {
-				this.i18n.changeLanguage(lng);
-			});
+			return this;
 		}
+
+		this.i18n.on('initialized', () => {
+			this.i18n.changeLanguage(lng);
+		});
 
 		return this;
 	}