+++ packages/livechat/CHANGELOG.md
@@ -1,5 +1,80 @@
 # @rocket.chat/livechat Change Log
 
+## 1.23.8-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+- @rocket.chat/gazzodown@22.0.0-rc.5
+  </details>
+
+## 1.23.5-rc.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/gazzodown@22.0.0-rc.4
+  </details>
+
+## 1.23.5-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/gazzodown@22.0.0-rc.3
+  </details>
+
+## 1.23.5-rc.2
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/gazzodown@22.0.0-rc.2
+  </details>
+
+## 1.23.5-rc.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/gazzodown@22.0.0-rc.1
+  </details>
+
+## 1.23.5-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/gazzodown@22.0.0-rc.0
+  </details>
+
+## 1.23.7
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+- @rocket.chat/gazzodown@21.0.3
+  </details>
+
+## 1.23.6
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+  - @rocket.chat/gazzodown@21.0.2
+  </details>
+
+## 1.23.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+- @rocket.chat/gazzodown@21.0.1
+</details>
+
 ## 1.23.4
 
 ### Patch Changes